web/satellite: fix invalid references to first onboarding step
This change amends invalid references to the first onboarding step.

Previously, the route for the first onboarding step was computed in a
router module using our old way of specifying frontend config values.
When we upgraded to our new config style, the module could no longer be
responsible for computing the route due to the module being independent
of the store holding the config. The route was removed, but some
references to it remained.

Change-Id: I8817174bb4acab44c5ff2624d6edb1bd3e675de2