Jenkinsfile.ui: Add UI Test jenkins job
This commit pushes the jenkinsfile that is required for a ui test to be
performed in the pipeline. Any branch with an assoicated github pr that
contains this file will automatically execute the jenkins job.

Further hooks will be implemented in a future PR to allow UI tests on
gerrit changes.